Refactor probabilistic model code and update example
Modified the existing code in probabilistic models file to improve the definition of the prior function. Added two new files for handling probabilistic distributions and nodes. Updated the normal likelihood example to fit with the new code structure. The changes also involved the removal of redundant imports and amending object values.