Update offsetX, offsetY, zoom, and id in tutorial.xircuits
OffsetX, offsetY, and zoom values have been adjusted to adapt to new layout design. Also, the 'id' of nodes has been altered. Removal of certain objects and parameter links in the code has led to a simplified version of the tutorials, enhancing the clarity and understandability for new users. The target ID and its corresponding target Port ID have been updated due to recent changes in the nodes connectivity.